+++ school-data-management-service-mapping/src/main/java/com/sdms/repo/TeacherRepository.java
@@ -6,5 +6,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface TeacherRepository extends JpaRepository<TeacherDetails, Long> {
+public interface TeacherRepository extends JpaRepository<TeacherDetails, Integer> {
 }
